Vivian Meazza:
I've reworked the Hunter model, attached, following observations that some
of the texture mapping didn't work with older cards. Melchior has tested it,
and apparently all is now OK. I've included a version with a crude droptank
simulation, I don't know what you might think about that.

I've also reduced the polygon count to 7000, not a great gain, but the model
looks better in places. I've also culled the vertices in the software - the
performance is now adequate at KSFO even on my Pentium III . The model is
in pretty good nick, but there is still much to be done on the systems
front.