Vivian Meazza:
The versions of the Hunter in CVS only display the parts of the aircraft visible from the cockpit in the interest of good frame rates.

When using multiplayer, the view is currently controlled by the receiving player, so only parts of the transmitting player\u2019s Hunter are displayed.

Pending a proper fix for this, I attach a couple of files which add hunter-mp to the inventory. This has the aircraft fully displayed in all views at the cost of a small penalty in frame rate.