Vivian Meazza:
The hunter-model file takes advantage of the Harald's new shader animation
to provide better modelling of the jet exhaust. I have also taken the
opportunity to fix up the shadows on the markings etc. It is no longer
necessary to select shadows - transparency to get the right effects. I've
also updated the pilot a little - added a flying helmet under the helmet,
and given him a more human face.